scripts/requirements.txt
pandas>=2.0.0
google-cloud-storage>=2.14.0
